// src/app/opmobilitybackend/models/get-movement-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GetExpoEventResponse } from '../models/get-expo-event-response';
import { GetProviderResponse } from '../models/get-provider-response';
import { GetWarehouseResponse } from '../models/get-warehouse-response';
import { RepairResponse } from '../models/repair-response';
export interface GetMovementResponse {

  /**
   * Destination data based on destKind
   */
  destData?: (GetWarehouseResponse | GetExpoEventResponse | GetProviderResponse | RepairResponse);
  destKind: 'Warehouse' | 'Event' | 'Provider' | 'Repair';
  destRefId: string;
  id: number;
  movedAt: string;
  notes?: string;

  /**
   * Origin data based on originKind
   */
  originData?: (GetWarehouseResponse | GetExpoEventResponse | GetProviderResponse | RepairResponse);
  originKind: 'Warehouse' | 'Event' | 'Provider' | 'Repair';
  originRefId: string;
  productId: number;
  type: 'Outbound' | 'Inbound';
  userId: number;
}