// src/app/opmobilitybackend/models/create-movement-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateMovementDto {

  /**
   * Destination kind (where the product is going to)
   */
  destKind: 'Warehouse' | 'Event' | 'Provider' | 'Repair';

  /**
   * Destination reference ID (ID of warehouse, event, or provider)
   */
  destRefId: number;

  /**
   * Date and time of movement
   */
  movedAt?: string;

  /**
   * Additional notes about this movement
   */
  notes?: string;

  /**
   * Origin kind (where the product is coming from)
   */
  originKind: 'Warehouse' | 'Event' | 'Provider' | 'Repair';

  /**
   * Origin reference ID (ID of warehouse, event, or provider)
   */
  originRefId: number;

  /**
   * The ID of the product being moved
   */
  productId: number;

  /**
   * The type of movement
   */
  type: 'Outbound' | 'Inbound';
}