// src/app/opmobilitybackend/models/get-expo-event-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GetTownResponseDto } from '../models/get-town-response-dto';
export interface GetExpoEventResponse {
  endDate: string;
  id: number;
  name: string;
  notes: string;
  ownerName: string;
  place: string;
  startDate: string;
  status: string;
  town?: GetTownResponseDto;
  townId: number;
}